lims-ai/lab_submission_rag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lab-submission-rag"
version = "1.0.0"
description = "Laboratory Submission RAG Service for TracSeq 2.0"
authors = [
    {name = "TracSeq Development Team", email = "dev@tracseq.com"},
]
readme = "README.md"
requires-python = ">=3.11"
keywords = ["rag", "laboratory", "nlp", "document-processing", "scientific"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core FastMCP and AI
    "fastmcp>=2.9.0",
    "anthropic>=0.34.0", 
    "openai>=1.3.0",
    "ollama>=0.3.0",
    
    # FastAPI and web components
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    
    # Data validation with email support
    "pydantic[email]>=2.0.0",
    "email-validator>=2.2.0",
    "pydantic-settings>=2.10.1",
    
    # HTTP client
    "httpx>=0.24.0",
    
    # Database integration
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    
    # Scientific computing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    
    # Document processing
    "pypdf>=4.0.0",
    "python-docx>=1.2.0",
    "langchain>=0.1.0",
    "chromadb>=0.4.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    
    # ML/AI dependencies  
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "scikit-learn>=1.3.0",
    
    # Testing
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0", 
    "pytest-cov>=6.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings>=0.23.0",
]

all = [
    "lab-submission-rag[dev,docs]"
]

[project.urls]
Homepage = "https://github.com/example/lab-submission-rag"
Documentation = "https://lab-submission-rag.readthedocs.io"
Repository = "https://github.com/example/lab-submission-rag.git"
Issues = "https://github.com/example/lab-submission-rag/issues"

[project.scripts]
lab-rag = "rag_orchestrator:main"

[tool.hatch.build.targets.wheel]
packages = ["rag", "models"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["anthropic.*", "openai.*", "langchain.*", "chromadb.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "rag: marks tests as RAG-specific",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["rag", "models"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 
